Managed to get the PyroprintSimulation new hotness running on my home machine with the following changes:
A lot of things.

But more specifically I simply cleaned up a bunch of code. renamed some
variables, fixed the broken ranges list (Not sure if it was broken on Mitch's
version on Tesla?) and did a whole lot of just... clean up. Code looks really
readable to me now (maybe no one else) and so I am committing! yay! tests on
Tesla to come!