Changed the name of pearson.cu to kernels.cu and removed pearson correlation kernels.
Pearson correlation kernels now have various versions in various files.
Constant memory version of the kernel is in const_pearson.cu, global memory
version is in ga_pearson.cu. Shared memory and textured memory versions will
eventually live in their own files accordingly.

I renamed the module that the kernel files get read into and then compiled so
that it is a bit more descriptive/appropriate.

finally fixed the issue with the global memory version, originally row and
column were switched (old problem) and then it took me forever to realize I was
passing the wrong number of arguments to the pearson kernel.

finally, the none option for memoryOpt has been changed to constantMem, to be
consistent with what the option entails, etc.