Created new branch to maintain the versions of Bob's old code for testing against. Haven't decided if I want to roll this into the mainstream branch or not so a new branch seems logical.
Primarily I referred to master branch and resurrected Bob's old kernel code in bob_pearson.cu, biogpu/correlation.py was modified to properly construct global memory for use in the bob_pearson.cu code. Additionally parseDNAfile was modified to have a flag for whether or not to use Bob's code. And if so it prepares things more similarly to how Bob's old code for 570 did. This includes enumerating all unique isolates on host.

runPyroSimTests was added here just for completeness sake.