Modified PyroprintSimulation to output different information when get_state is called.
Modified Sequence.py to use allele as identifying instead of sequence (this way
unique sequences are based on unique alleles instead of unique sequences which
may have duplicate alleles).

Additionally, SequenceParser was modified while I ensured it was correct, and I
added a new output for when 'STATS' is an environment variable that outputs
statistics about the alleles being pyroprinted and run through the simulation.

Finally, I modified config.cfg to be a bit more structured and clear as to what
dispensation and sequencing primers are for what regions.